# [Feature] Optimize search migration [#OSF-9037]