--- a/tasks/__init__.py
+++ b/tasks/__init__.py
@@ -279,7 +279,7 @@ def elasticsearch(ctx):
         print('Your system is not recognized, you will have to start elasticsearch manually')
 
 @task
-def migrate_search(ctx, delete=False, index=settings.ELASTIC_INDEX):
+def migrate_search(ctx, delete=True, remove=False, index=settings.ELASTIC_INDEX):
     """Migrate the search-enabled models."""
     from website.app import init_app
     init_app(routes=False, set_backends=False)
@@ -291,7 +291,7 @@ def migrate_search(ctx, delete=False, index=settings.ELASTIC_INDEX):
     for logger in SILENT_LOGGERS:
         logging.getLogger(logger).setLevel(logging.ERROR)
 
-    migrate(delete, index=index)
+    migrate(delete, remove=remove, index=index)
 
 
 @task
@@ -317,7 +317,7 @@ def rebuild_search(ctx):
     print('Creating index {}'.format(settings.ELASTIC_INDEX))
     print('----- PUT {}'.format(url))
     requests.put(url)
-    migrate_search(ctx)
+    migrate_search(ctx, delete=False)
 
 
 @task