--- a/website/search_migration/migrate.py
+++ b/website/search_migration/migrate.py
@@ -11,7 +11,10 @@ from elasticsearch import helpers
 
 import website.search.search as search
 from website.search.elastic_search import client
-from website.search_migration import JSON_UPDATE_NODES_SQL, JSON_UPDATE_FILES_SQL, JSON_UPDATE_USERS_SQL
+from website.search_migration import (
+    JSON_UPDATE_NODES_SQL, JSON_DELETE_NODES_SQL,
+    JSON_UPDATE_FILES_SQL, JSON_DELETE_FILES_SQL,
+    JSON_UPDATE_USERS_SQL, JSON_DELETE_USERS_SQL)
 from scripts import utils as script_utils
 from osf.models import OSFUser, Institution, AbstractNode, BaseFileNode
 from website import settings
@@ -21,17 +24,20 @@ from website.search.search import update_institution
 
 logger = logging.getLogger(__name__)
 
-def sql_migrate(index, sql, max_id, increment, **kwargs):
+def sql_migrate(index, sql, max_id, increment, es_args=None, **kwargs):
     """ Run provided SQL and send output to elastic.
 
     :param str index: Elastic index to update (formatted into `sql`)
     :param str sql: SQL to format and run. See __init__.py in this module
     :param int max_id: Last known object id. Indicates when to stop paging
     :param int increment: Page size
+    :param  dict es_args:  Dict or None, to pass to `helpers.bulk`
     :kwargs: Additional format arguments for `sql` arg
 
     :return int: Number of migrated objects
     """
+    if es_args is None:
+        es_args = {}
     total_pages = int(ceil(max_id / float(increment)))
     total_objs = 0
     page_start = 0
@@ -56,11 +62,11 @@ def sql_migrate(index, sql, max_id, increment, **kwargs):
             ser_objs = cursor.fetchone()[0]
             if ser_objs:
                 total_objs += len(ser_objs)
-                helpers.bulk(client(), ser_objs)
+                helpers.bulk(client(), ser_objs, **es_args)
         page_start = page_end
     return total_objs
 
-def migrate_nodes(index, increment=10000):
+def migrate_nodes(index, delete, increment=10000):
     logger.info('Migrating nodes to index: {}'.format(index))
     max_nid = AbstractNode.objects.last().id
     total_nodes = sql_migrate(
@@ -70,8 +76,19 @@ def migrate_nodes(index, increment=10000):
         increment,
         spam_flagged_removed_from_search=settings.SPAM_FLAGGED_REMOVE_FROM_SEARCH)
     logger.info('{} nodes migrated'.format(total_nodes))
+    if delete:
+        logger.info('Preparing to delete old node documents')
+        max_nid = AbstractNode.objects.last().id
+        total_nodes = sql_migrate(
+            index,
+            JSON_DELETE_NODES_SQL,
+            max_nid,
+            increment,
+            es_args={'raise_on_error': False},  # ignore 404s
+            spam_flagged_removed_from_search=settings.SPAM_FLAGGED_REMOVE_FROM_SEARCH)
+        logger.info('{} nodes marked deleted'.format(total_nodes))
 
-def migrate_files(index, increment=10000):
+def migrate_files(index, delete, increment=10000):
     logger.info('Migrating files to index: {}'.format(index))
     max_fid = BaseFileNode.objects.last().id
     total_files = sql_migrate(
@@ -81,9 +98,19 @@ def migrate_files(index, increment=10000):
         increment,
         spam_flagged_removed_from_search=settings.SPAM_FLAGGED_REMOVE_FROM_SEARCH)
     logger.info('{} files migrated'.format(total_files))
+    if delete:
+        logger.info('Preparing to delete old file documents')
+        max_fid = BaseFileNode.objects.last().id
+        total_files = sql_migrate(
+            index,
+            JSON_DELETE_FILES_SQL,
+            max_fid,
+            increment,
+            es_args={'raise_on_error': False},  # ignore 404s
+            spam_flagged_removed_from_search=settings.SPAM_FLAGGED_REMOVE_FROM_SEARCH)
+        logger.info('{} files marked deleted'.format(total_files))
 
-
-def migrate_users(index, increment=10000):
+def migrate_users(index, delete, increment=10000):
     logger.info('Migrating users to index: {}'.format(index))
     max_uid = OSFUser.objects.last().id
     total_users = sql_migrate(
@@ -92,13 +119,29 @@ def migrate_users(index, increment=10000):
         max_uid,
         increment)
     logger.info('{} users migrated'.format(total_users))
-
+    if delete:
+        logger.info('Preparing to delete old user documents')
+        max_uid = OSFUser.objects.last().id
+        total_users = sql_migrate(
+            index,
+            JSON_DELETE_USERS_SQL,
+            max_uid,
+            increment,
+            es_args={'raise_on_error': False})  # ignore 404s
+        logger.info('{} users marked deleted'.format(total_users))
 
 def migrate_institutions(index):
     for inst in Institution.objects.filter(is_deleted=False):
         update_institution(inst, index)
 
-def migrate(delete, index=None, app=None):
+def migrate(delete, remove=False, index=None, app=None):
+    """Reindexes relevant documents in ES
+
+    :param bool delete: Delete documents that should not be indexed
+    :param bool remove: Removes old index after migrating
+    :param str index: index alias to version and migrate
+    :param App app: Flask app for context
+    """
     index = index or settings.ELASTIC_INDEX
     app = app or init_app('website.settings', set_backends=True, routes=True)
 
@@ -113,14 +156,14 @@ def migrate(delete, index=None, app=None):
 
     if settings.ENABLE_INSTITUTIONS:
         migrate_institutions(new_index)
-    migrate_nodes(new_index)
-    migrate_files(new_index)
-    migrate_users(new_index)
+    migrate_nodes(new_index, delete=delete)
+    migrate_files(new_index, delete=delete)
+    migrate_users(new_index, delete=delete)
 
     set_up_alias(index, new_index)
 
-    if delete:
-        delete_old(new_index)
+    if remove:
+        remove_old_index(new_index)
 
     ctx.pop()
 
@@ -155,7 +198,7 @@ def set_up_alias(old_index, index):
     es_client().indices.put_alias(index=index, name=old_index)
 
 
-def delete_old(index):
+def remove_old_index(index):
     old_version = int(index.split('_v')[1]) - 1
     if old_version < 1:
         logger.info('No index before {} to delete'.format(index))