@mfraezz 2018-01-09
Affirmative, @caseyrollins :shipit: 

Only thing we discussed  that I haven't done is update the scratchpad, as I've been tied up investigating staging slowness since finishing this on Friday.