@sloria 2018-01-09
👍  Thanks for doing the in-person CR Matt and Casey. It's comforting to know that > 1 person will be able to read this. =)

Merged into `release/next-next-djangole`