@sloria 2018-01-10
After discussing with Nici, we decided it'd be better to get this in sooner, so I'm going to merge this into develop so that it can be QA'd on staging.